Changed redux-subspace-loop to only export a single function
